refactor(configs): update application-local.yml and move config files
