+++ src/main/kotlin/com/jlima/bookstoremanager/config/openapi/SwaggerConfig.kt
@@ -1,4 +1,4 @@
-package com.jlima.bookstoremanager.config
+package com.jlima.bookstoremanager.config.openapi
 
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
