+++ src/main/resources/application-local.yml
@@ -1,3 +1,7 @@
 logging:
   level:
     com.jlima: DEBUG
+spring:
+  jpa:
+    show-sql: true
+    generate-ddl: true
